refactor: separate webui api config
