+++ webui/react/src/pages/Dashboard.tsx
@@ -14,7 +14,8 @@ import Users from 'contexts/Users';
 import usePolling from 'hooks/usePolling';
 import { useRestApiSimple } from 'hooks/useRestApi';
 import useStorage from 'hooks/useStorage';
-import { ExperimentsParams, getExperimentSummaries } from 'services/api';
+import { getExperimentSummaries } from 'services/api';
+import { ExperimentsParams } from 'services/types';
 import { ShirtSize } from 'themes';
 import {
   Command, CommandState, Experiment, RecentTask, ResourceType, RunState, TaskType,
