+++ webui/react/src/pages/Determined.tsx
@@ -11,11 +11,12 @@ import usePolling from 'hooks/usePolling';
 import useRestApi, { useRestApiSimple } from 'hooks/useRestApi';
 import { ioAgents, ioGenericCommands, ioUsers } from 'ioTypes';
 import { detRoutes } from 'routes';
-import { ExperimentsParams, getExperimentSummaries } from 'services/api';
+import { getExperimentSummaries } from 'services/api';
 import {
   jsonToAgents, jsonToCommands, jsonToNotebooks,
   jsonToShells, jsonToTensorboards, jsonToUsers,
 } from 'services/decoder';
+import { ExperimentsParams } from 'services/types';
 import { Agent, Command, Experiment, RunState, User } from 'types';
 
 import css from './Determined.module.scss';
