+++ webui/react/src/services/types.ts
@@ -1,0 +1,19 @@
+import { CommandType } from 'types';
+
+export interface ExperimentsParams {
+  states?: string[];
+}
+
+export interface KillExpParams {
+  experimentId: number;
+}
+
+export interface KillCommandParams {
+  commandId: string;
+  commandType: CommandType;
+}
+
+export interface PatchExperimentParams {
+  experimentId: number;
+  body: Record<keyof unknown, unknown> | string;
+}
